validate delta methods
