+++ src/atc/delta/delta_handle.py
@@ -2,25 +2,12 @@ from typing import List, Union
 
 from pyspark.sql import DataFrame
 
-from atc.atc_exceptions import AtcException
 from atc.configurator.configurator import Configurator
 from atc.functions import get_unique_tempview_name, init_dbutils
 from atc.spark import Spark
 from atc.tables.SparkHandle import SparkHandle
 from atc.utils.CheckDfMerge import CheckDfMerge
 from atc.utils.GetMergeStatement import GetMergeStatement
-
-
-class DeltaHandleException(AtcException):
-    pass
-
-
-class DeltaHandleInvalidName(DeltaHandleException):
-    pass
-
-
-class DeltaHandleInvalidFormat(DeltaHandleException):
-    pass
 
 
 class DeltaHandle(SparkHandle):
@@ -38,29 +25,6 @@ class DeltaHandle(SparkHandle):
             location=tc.table_property(id, "path", ""),
             data_format=tc.table_property(id, "format", "delta"),
         )
-
-    def _validate(self):
-        """Validates that the name is either db.table or just table."""
-        if not self._name:
-            if not self._location:
-                raise DeltaHandleInvalidName(
-                    "Cannot create DeltaHandle without name or path"
-                )
-            self._name = f"delta.`{self._location}`"
-        else:
-            name_parts = self._name.split(".")
-            if len(name_parts) == 1:
-                self._db = None
-                self._table_name = name_parts[0]
-            elif len(name_parts) == 2:
-                self._db = name_parts[0]
-                self._table_name = name_parts[1]
-            else:
-                raise DeltaHandleInvalidName(f"Could not parse name {self._name}")
-
-        # only format delta is supported.
-        if self._data_format != "delta":
-            raise DeltaHandleInvalidFormat("Only format delta is supported.")
 
     def read(self) -> DataFrame:
         """Read table by path if location is given, otherwise from name."""
